+++ CHANGELOG.md
@@ -5,10 +5,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+[1.3.3]
+### Fixed
+- Incorrect functions generated when providing both `suffix: false` and other  options to the `resource` macro
+
 [1.3.2]
 ------------
 ### Fixed
-- Incorrect function name generation when providing both `suffix: false` and other options to the `resource` macroo
+- Incorrect function name generation when providing both `suffix: false` and other options to the `resource` macro
 
 [1.3.1]
 ------------